move fetch data in bookingForm component
