+++ src/components/ApartmentCards.js
@@ -1,8 +1,7 @@
-import { React, useEffect, useState } from 'react';
+import { React } from 'react';
 import ApartmentCard from './ApartmentCard';
 import { Grid } from '@material-ui/core';
 import { makeStyles } from '@material-ui/styles';
-import getApartments from '../services/apartmentsSearch';
 import useStore from '../store';
 
 const useStyles = makeStyles({
@@ -13,19 +12,9 @@ const useStyles = makeStyles({
 });
 
 const ApartmentCards = () => {
-  const { location } = useStore((state) => state.bookingSearch);
-
-  const [apartmentsData, setApartmentsData] = useState([]);
-
-  useEffect(() => {
-    const fetchApartments = async () => {
-      const response = await getApartments(location);
-      setApartmentsData(response.data.apartments);
-    };
-    fetchApartments();
-  }, [location]);
-
   const classes = useStyles();
+
+  const apartmentsData = useStore((state) => state.apartmentsList);
   return (
     /* List of cards (colymn) */
     <Grid container direction="colymn" className={classes.cards}>
