+++ src/components/Booking.js
@@ -3,6 +3,7 @@ import { makeStyles, Grid } from '@material-ui/core';
 import BookingForm from './BookingForm';
 import BookingResults from './BookingResults';
 import BookingFilters from './BookingFilters';
+import useStore from '../store';
 
 const useStyles = makeStyles({
   root: {
@@ -17,20 +18,25 @@ const useStyles = makeStyles({
 
 const Booking = () => {
   const classes = useStyles();
+
+  const apartmentsList = useStore((state) => state.apartmentsList);
+
   return (
     <>
       <div className={classes.root}>
         <h2 style={{ padding: 0 }}>Book your stay</h2>
         <BookingForm />
       </div>
-      <Grid container spacing={4} className={classes.results}>
-        <Grid item md={3} xs={12}>
-          <BookingFilters />
+      {apartmentsList.length !== 0 && (
+        <Grid container spacing={4} className={classes.results}>
+          <Grid item md={3} xs={12}>
+            <BookingFilters />
+          </Grid>
+          <Grid item md={9} xs={12}>
+            <BookingResults />
+          </Grid>
         </Grid>
-        <Grid item md={9} xs={12}>
-          <BookingResults />
-        </Grid>
-      </Grid>
+      )}
     </>
   );
 };
